Moved `TextStyle` functions to `DesignTokens`
