+++ crates/re_viewer/src/ui/welcome_screen/no_data_ui.rs
@@ -11,7 +11,7 @@ pub fn no_data_ui(ui: &mut egui::Ui) {
                 egui::RichText::new(super::welcome_section::WELCOME_SCREEN_TITLE)
                     .weak()
                     .line_height(Some(36.0))
-                    .text_style(re_ui::ReUi::welcome_screen_h2()),
+                    .text_style(re_ui::DesignTokens::welcome_screen_h2()),
             )
             .wrap(),
         );
@@ -27,7 +27,7 @@ pub fn no_data_ui(ui: &mut egui::Ui) {
                     egui::Label::new(
                         egui::RichText::new(text)
                             .color(ui.visuals().weak_text_color())
-                            .text_style(re_ui::ReUi::welcome_screen_body()),
+                            .text_style(re_ui::DesignTokens::welcome_screen_body()),
                     )
                     .wrap(),
                 );
@@ -44,7 +44,7 @@ pub fn no_data_ui(ui: &mut egui::Ui) {
             .button(
                 egui::RichText::new("Go to documentation →")
                     .weak()
-                    .text_style(re_ui::ReUi::welcome_screen_body()),
+                    .text_style(re_ui::DesignTokens::welcome_screen_body()),
             )
             .on_hover_cursor(egui::CursorIcon::PointingHand)
             .clicked()
